pyproject.toml
[project]
name = "wallet-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2",
    "django-cors-headers>=4.7.0",
    "django-unfold>=0.54.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "gunicorn>=23.0.0",
    "psycopg>=3.2.6",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=5.1.0",
    "django-environ>=0.12.0",
    "django-silk>=5.3.2",
    "drf-extra-fields>=3.7.0",
    "pillow>=11.1.0",
    "ruff>=0.11.4",
]
